fixed leak possibility

+++ ft_strlcpy.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strlcpy.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ebinjama <ebinjama@student.42.fr>          +#+  +:+       +#+        */
+/*   By: ebinjama <ebinjama@student.42abudhabi.ae>  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/04 17:59:52 by ebinjama          #+#    #+#             */
-/*   Updated: 2023/11/07 19:15:22 by ebinjama         ###   ########.fr       */
+/*   Updated: 2023/11/09 05:48:16 by ebinjama         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,13 +14,13 @@
 
 size_t	ft_strlcpy(char *dst, const char *src, size_t sz)
 {
-	int	i;
-	int	src_len;
+	int		i;
+	size_t	src_len;
 
 	if (!sz)
 		return (ft_strlen(src));
 	src_len = ft_strlen(src);
-	if ((size_t)src_len > --sz)
+	if (src_len > --sz)
 		src_len = sz;
 	i = -1;
 	while (++i < src_len)
